backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
PyPDF2>=3.0.0
python-docx>=1.1.0
sentence-transformers>=2.7.0
pandas>=2.2.0
scikit-learn>=1.4.0
pydantic>=2.6.0
python-dotenv>=1.0.0
